requirements.txt
websockets==12.0
paho-mqtt==2.1.0
